Moved test_cookies_jar to asgi test package
